+++ src/drawing.c
@@ -6,29 +6,29 @@
 /*   By: levensta <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/29 22:34:56 by levensta          #+#    #+#             */
-/*   Updated: 2021/02/16 20:44:01 by levensta         ###   ########.fr       */
+/*   Updated: 2021/02/17 23:35:41 by levensta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-static int	count_column(float x, float y, t_all *cub)
+static int	count_column(double x, double y, t_all *cub)
 {
 	int		column_h;
 
-	cub->dists[cub->x] = sqrtf(powf(x, 2) + powf(y, 2));
-	cub->dists[cub->x] *= cosf(fabsf(cub->plr.route * 360.0f - cub->ray \
+	cub->dists[cub->x] = sqrt(powf(x, 2) + powf(y, 2));
+	cub->dists[cub->x] *= cos(fabs(cub->plr.route * 360.0f - cub->ray \
 	* 360.0f) * (M_PI / 180.0f));
 	column_h = cub->s_width / 2;
-	column_h = (float)column_h / tanf((FOV * 360.0f / 2.0f) * (M_PI / 180.0f));
-	column_h = (int)ceilf((float)column_h / cub->dists[cub->x]);
+	column_h = (double)column_h / tan((FOV * 360.0f / 2.0f) * (M_PI / 180.0f));
+	column_h = (int)ceil((double)column_h / cub->dists[cub->x]);
 	return (column_h);
 }
 
 void		draw_ceil(t_all *cub, int column_h)
 {
 	int		i;
-	float	j;
+	double	j;
 
 	if (column_h < 0)
 		column_h = 0;
@@ -49,10 +49,10 @@ void		draw_floor(t_all *cub, int column_h)
 		my_mlx_pixel_put(cub, cub->x, j, (unsigned int)cub->scene.flooring);
 }
 
-void		draw_texture(t_all *cub, float hit, int size, int n)
+void		draw_texture(t_all *cub, double hit, int size, int n)
 {
 	unsigned int	color;
-	float			j;
+	double			j;
 	int				i;
 	int				max;
 
@@ -61,7 +61,7 @@ void		draw_texture(t_all *cub, float hit, int size, int n)
 	max = (cub->s_height + size) / 2;
 	if (i < 0)
 	{
-		j = ((float)cub->txt[n].height / (float)size) * (-i);
+		j = ((double)cub->txt[n].height / (double)size) * (-i);
 		i = 0;
 		max = cub->s_height;
 	}
@@ -72,7 +72,7 @@ void		draw_texture(t_all *cub, float hit, int size, int n)
 		+ (int)hit * (cub->txt[n].bpp / 8)));
 		my_mlx_pixel_put(cub, cub->x, i, color);
 		i++;
-		j += (float)cub->txt[n].height / (float)size;
+		j += (double)cub->txt[n].height / (double)size;
 	}
 }
 
@@ -85,15 +85,15 @@ void		drawing_room(t_all *cub, char is_x)
 	if (is_x)
 	{
 		if (cub->ray >= 0 && cub->ray < 0.5f)
-			draw_texture(cub, cub->y1 - floorf(cub->y1), cub->column_h, 2);
+			draw_texture(cub, cub->y1 - floor(cub->y1), cub->column_h, 2);
 		else
-			draw_texture(cub, ceilf(cub->y1) - cub->y1, cub->column_h, 3);
+			draw_texture(cub, ceil(cub->y1) - cub->y1, cub->column_h, 3);
 	}
 	else
 	{
 		if (cub->ray >= 0.25f && cub->ray < 0.75f)
-			draw_texture(cub, ceilf(cub->x1) - cub->x1, cub->column_h, 0);
+			draw_texture(cub, ceil(cub->x1) - cub->x1, cub->column_h, 0);
 		else
-			draw_texture(cub, cub->x1 - floorf(cub->x1), cub->column_h, 1);
+			draw_texture(cub, cub->x1 - floor(cub->x1), cub->column_h, 1);
 	}
 }
